refactor: changed folder name in org.apache.seata.server.storage.raft.sore from sore to store (#7236)
