+++ changes/en-us/2.x.md
@@ -86,6 +86,7 @@ Add changes here for all PR submitted to the 2.x branch.
 ### refactor:
 
 - [[#7145](https://github.com/apache/incubator-seata/pull/7145)] refactor the code that does not comply with license requirements
+- [[#7236](https://github.com/apache/incubator-seata/pull/7236)] changed folder name in org.apache.seata.server.storage.raft.sore from sore to store
 
 
 ### doc:
@@ -115,5 +116,6 @@ Thanks to these contributors for their code commits. Please report an unintended
 - [s-ramyalakshmi](https://github.com/s-ramyalakshmi)
 - [YoWuwuuuw](https://github.com/YoWuwuuuw)
 - [mehedikhan72](https://github.com/mehedikhan72)
+- [AndrewSf](https://github.com/andrewseif)
 
 Also, we receive many valuable issues, questions and advices from our community. Thanks for you all.
