+++ changes/zh-cn/2.x.md
@@ -85,6 +85,7 @@
 ### refactor:
 
 - [[#7145](https://github.com/apache/incubator-seata/pull/7145)] 重构不满足 license 要求的代码
+- [[#7236](https://github.com/apache/incubator-seata/pull/7236)] 将 org.apache.seata.server.storage.raft.sore 中的文件夹名称从 sore 更改为 store
 
 非常感谢以下 contributors 的代码贡献。若有无意遗漏，请报告。
 
@@ -108,5 +109,6 @@
 - [iAmClever](https://github.com/iAmClever)
 - [s-ramyalakshmi](https://github.com/s-ramyalakshmi)
 - [YoWuwuuuw](https://github.com/YoWuwuuuw)
+- [AndrewSf](https://github.com/andrewseif)
 
 同时，我们收到了社区反馈的很多有价值的issue和建议，非常感谢大家。
