+++ server/src/main/resources/META-INF/services/org.apache.seata.server.store.VGroupMappingStoreManager
@@ -17,4 +17,4 @@
 org.apache.seata.server.storage.db.store.DataBaseVGroupMappingStoreManager
 org.apache.seata.server.storage.file.store.FileVGroupMappingStoreManager
 org.apache.seata.server.storage.redis.store.RedisVGroupMappingStoreManager
-org.apache.seata.server.storage.raft.sore.RaftVGroupMappingStoreManager
+org.apache.seata.server.storage.raft.store.RaftVGroupMappingStoreManager
